[#21] chore : keywords 더미데이터 분류
