+++ src/components/SideNavbar/KeywordDropdown.js
@@ -1,35 +1,8 @@
 import React from 'react';
 import * as S from 'components/SideNavbar/styles';
+import keywords from './dummy';
 
 const KeywordDropdown = () => {
-  const keywords = [
-    {
-      id: 1,
-      text: '김이정',
-      update: 10,
-    },
-    {
-      id: 2,
-      text: '이유정',
-      update: 10,
-    },
-    {
-      id: 3,
-      text: '국가장학금',
-      update: 10,
-    },
-    {
-      id: 4,
-      text: '근로장학생',
-      update: 10,
-    },
-    {
-      id: 5,
-      text: '성적',
-      update: 10,
-    },
-  ];
-
   return (
     <S.KeywordDropdown>
       <S.KeywordMain>키워드</S.KeywordMain>
