+++ src/components/SideNavbar/dummy.js
@@ -1,0 +1,29 @@
+const keywords = [
+  {
+    id: 1,
+    text: '김이정',
+    update: 10,
+  },
+  {
+    id: 2,
+    text: '이유정',
+    update: 10,
+  },
+  {
+    id: 3,
+    text: '국가장학금',
+    update: 10,
+  },
+  {
+    id: 4,
+    text: '근로장학생',
+    update: 10,
+  },
+  {
+    id: 5,
+    text: '성적',
+    update: 10,
+  },
+];
+
+export default keywords;
